rename script to sync with animations
